requirements.txt
aiosmtpd~=1.2.4
aiosmtplib~=1.1.4
aiohttp~=3.7.3
dnspython~=2.1.0
ezt~=1.1
